Add reused link_to_document helper to Bops Core
