+++ engines/bops_admin/app/helpers/bops_admin/application_helper.rb
@@ -71,16 +71,5 @@ module BopsAdmin
         {text: "Settings", href: profile_path, active: active_page_key?("settings")}
       ]
     end
-
-    def link_to_document(link_text, document, **args)
-      new_tab = /(new (window|tab)|<img\b)/.match?(link_text) ? "" : true
-
-      govuk_link_to(
-        link_text,
-        main_app.uploaded_file_url(document.blob),
-        new_tab:,
-        **args
-      )
-    end
   end
 end
