+++ engines/bops_core/app/helpers/bops_core/application_helper.rb
@@ -39,5 +39,24 @@ module BopsCore
 
       super
     end
+
+    def link_to_document(link_text, document, **args)
+      new_tab = /(new (window|tab)|<img\b)/.match?(link_text) ? "" : true
+
+      govuk_link_to(
+        link_text,
+        url_for_document(document),
+        new_tab:,
+        **args
+      )
+    end
+
+    def url_for_document(document)
+      if document.published?
+        main_app.api_v1_planning_application_document_url(document.planning_application, document)
+      else
+        main_app.uploaded_file_url(document.blob)
+      end
+    end
   end
 end
