+++ engines/bops_reports/app/helpers/bops_reports/application_helper.rb
@@ -21,24 +21,5 @@ module BopsReports
     def map_link(full_address)
       "https://google.co.uk/maps/place/#{CGI.escape(full_address)}"
     end
-
-    def link_to_document(link_text, document, **args)
-      new_tab = /(new (window|tab)|<img\b)/.match?(link_text) ? "" : true
-
-      govuk_link_to(
-        link_text,
-        url_for_document(document),
-        new_tab:,
-        **args
-      )
-    end
-
-    def url_for_document(document)
-      if document.published?
-        main_app.api_v1_planning_application_document_url(document.planning_application, document)
-      else
-        main_app.uploaded_file_url(document.blob)
-      end
-    end
   end
 end
